Upgrade to React 19, MUI 7, and migrate from CRA to Vite
Major dependency upgrades:
- React 18 → 19, react-dom 18 → 19
- MUI Material & Icons 5 → 7 (migrate Grid to v2 API)
- @testing-library/react 13 → 16, user-event 13 → 14, jest-dom 5 → 6
- web-vitals 2 → 5
- Remove unused react-router-dom

Build system migration (CRA → Vite):
- Replace react-scripts with vite + @vitejs/plugin-react
- Replace Jest with Vitest for testing
- Replace CRA SVG imports with vite-plugin-svgr
- Move index.html to root with ES module entry point
- Rename .js → .jsx for all JSX files
- Remove CRA boilerplate (reportWebVitals, theme.js, App.css)
- Dependencies: 1347 → 246 packages, 9 → 0 vulnerabilities
