--- src/theme.js
@@ -1,11 +1,0 @@
-// src/theme.js
-
-import { createTheme } from '@mui/material/styles';
-
-const theme = createTheme({
-    typography: {
-        fontFamily: '"Fira Code", monospace',
-    },
-  });
-
-export default theme;
